SARIF: Restructure handling
* Add automationDetails
* Record started-at
* upload-sarif: Call github /code-scanning/sarifs api directly

Replace github/codeql-action/upload-sarif with a simpler api call.

- Avoid `/usr/bin/gh: Argument list too long` by using curl w/ `-d @`

(`gh`` appears to support `@` as well, but there doesn't seem to be much
reason to use it.)

* Change parse_warnings to return a reference

* Support creating multiple location reports

Unfortunately, GitHub's current implementation says:
https://github.com/github/docs/blob/b68cb60117832e1b3d4b85a011b36745caecde23/content/code-security/code-scanning/integrating-with-code-scanning/sarif-support-for-code-scanning.md?plain=1#L196C515-L196C558
> Only the first value of this array is used.

So, while this code would be nice, it is not active.

* t/Sarif.t -- use self-references

* draft implementation of partialFingerprint

Based on a hash of 100 characters before the current line,
unless the current line is the first line in which case it's a hash of
the first line.

- Whitespace runs are normalized to ` `.
- Hash collisions are given a `:hit` suffix to disambiguate.

* Replace `jd` with `Test::More::is_deeply`